server/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-cors==6.0.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
nvidia-ml-py3==7.352.0
paho-mqtt==2.1.0
psutil==7.1.0
scheduler==0.8.8
typeguard==4.4.4
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
